initial import of basic psiTurk files
basically all psiTurk projects have a static/ and a templates/ folder.  these hold the javascript, css, and html files for the experiment.

in addition there is a config.txt file which has your configuration options.  custom.py is an option a file that lets you write code to automatically compute bonuses and stuff.  for this project it is probably not necessary so I will delete it in the next commit.

you can create these files anytime by typing 'psiturk-setup-example' when you have it installed.